+++ tests/fixture/remove_export_exprs/preserveData/remove-default-decl/output.js
@@ -1,0 +1,11 @@
+import fs from 'fs';
+import other from 'other';
+const [a, b, ...rest] = fs.promises;
+const [foo, bar] = other;
+export async function getData() {
+    a;
+    b;
+    rest;
+    bar;
+}
+export default function() {}